rev commit

+++ app/src/main/java/com/example/themarketer/data/model/MenuSections/Price.kt
@@ -5,9 +5,9 @@ import com.google.gson.annotations.SerializedName
 
 data class Price(
     @SerializedName("price")
-    var price: Int,
+    var price: Float,
     @SerializedName("price_after_diccount")
-    var priceAfterDiccount: Int,
+    var priceAfterDiccount: Float,
     @SerializedName("type")
     var type: String
 )
